pyproject.toml
[project]
name = "pokemon-card-scanner"
version = "1.0.0"
description = "Production-ready Pokemon card scanner using Gemini AI and Pokemon TCG API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "google-generativeai>=0.8.0",
    "httpx>=0.28.1",
    "opencv-python-headless>=4.10.0",
    "pillow>=11.2.1",
    "pillow-heif>=0.18.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.0.0",
    "python-json-logger>=3.2.1",
    "tenacity>=9.0.0",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.16.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.12.0",
    "responses>=0.22.0",
    "factory-boy>=3.3.0",
    "ruff>=0.11.12",
]
testing = [
    "aiohttp>=3.9.0",
    "jinja2>=3.1.0",
    "plotly>=5.17.0",
    "pandas>=2.1.0",
    "rich>=13.7.0",
]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
extend-select = ["E", "F", "W", "I", "N", "UP"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
pythonpath = ["."]
addopts = "--cov=src --cov-report=html --cov-report=term-missing --cov-fail-under=80"
asyncio_mode = "auto"
